Update script to support external DUI resources
Currently, the script cannot handle external resources that will eventually go into SystemUI, so they will not be part of this repo. This is a quick and dirty way to copy all of DUI resources after they're synced and put them in SystemUI, so theme devs can theme the extra resources in for their final place.

Theres probably a better way to do it, but it works for now.

Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>